Revert "Move to C++17. Remove Mapbox Variant and Any/Optional libs (which were"
This reverts commit 3e1a4c0e5705db8943217bab85ebea9067e9558f.

+++ src/dawn/renderer/SceneGraph.cpp
@@ -86,7 +86,7 @@ void SceneGraph::updateSceneGraph() {
         // TODO: Culling?
         Renderable* renderable = node->data.renderable.get();
         if (renderable) {
-            for (int c = 0; c < cameras.size(); ++c) {
+            for (usize c = 0; c < cameras.size(); ++c) {
                 int f = frame_to_frame_id.at(cameras[c].scene_node->frame());
                 Mat4& model_matrix = system_model_matrices_per_frame_[f][node];
                 render_operations_per_camera_[c].emplace_back(
